Tests/LibWeb/Text/expected/css/unicode-bidi-computed-values.txt

normal
embed
isolate
bidi-override
isolate-override
plaintext
normal
normal
normal
normal
